app/backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
reportlab==4.2.5
aiofiles==24.1.0
psycopg2-binary==2.9.9
requests==2.32.3
